veracheck/backend/requirements.txt
# VeraCheck Backend Dependencies

# Web framework
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pydantic==2.7.1
pydantic-settings==2.3.0

# ML / Embeddings
sentence-transformers==2.7.0
torch==2.3.0
transformers==4.41.0
tokenizers==0.19.1
accelerate==0.30.0

# FAISS (vector search)
faiss-cpu==1.8.0

# Language detection
langdetect==1.0.9

# Redis
redis[asyncio]==5.0.4

# Utilities
numpy==1.26.4
httpx==0.27.0
python-dotenv==1.0.1
deep-translator==1.11.4
sentencepiece>=0.1.99

# Dev
pytest==8.2.0
pytest-asyncio==0.23.7
